+++ tempest_fastapi_sdk/__init__.py
@@ -359,7 +359,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.88.0"
+__version__: str = "0.89.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",